+++ lib/dart_ui_stub/dart_ui.dart
@@ -1,0 +1,2 @@
+// stub dart_ui for ui.platformViewRegistry exposition on compile time
+export 'dart_ui_fake.dart' if (dart.library.html) 'dart_ui_real.dart';