+++ lib/dart_ui_stub/dart_ui_fake.dart
@@ -1,0 +1,7 @@
+import 'dart:html' as html;
+
+// stub dart_ui for ui.platformViewRegistry exposition on compile time
+class platformViewRegistry {
+  static void registerViewFactory(
+      String viewTypeId, html.Element Function(int viewId) viewFactory) {}
+}