+++ lib/dart_ui_stub/dart_ui_real.dart
@@ -1,0 +1,2 @@
+// stub dart_ui for ui.platformViewRegistry exposition on compile time
+export 'dart:ui';