+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add missing documentation
 * Fix analysis errors
+* Add dart_ui stub to be compliant with static analysis on compile time
 
 ## 1.0.2